update server file for database branch
